+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More descriptive manifest validation error messages
 - Add `default_namespace`, `registry_url` parameters to `compote registry login` command
 - Create a profile in `idf_component_manager.yml` if it does not exist when executing `compote registry login` command
+- Option to check for new versions of components each time CMake is triggered (IDF_COMPONENT_CHECK_NEW_VERSION env variable)
 
 ### Fixed